Mortgage Calculator
This spreadsheet was originally created by Chris Gamlin in 2019, this is a lighter version of the spreadsheet I have adapted to make it easier to get an idea of what you will be paying over the loan repayment period, with just the loan interest rate and the number of years over which you plan to pay off the mortgage. 

The original version can be found here:

https://forums.moneysavingexpert.com/discussion/1157173/my-excel-mortgage-spreadsheet

On the money-saving expert forum.